do not remove secrets from the config when they are saved
