+++ client/src/features/project/utils/projectCloudStorage.utils.ts
@@ -416,6 +416,23 @@ export function findSensitive(
     : [];
 }
 
+export function storageDefinitionAfterSavingCredentialsFromConfig(
+  cs: SessionStartCloudStorageConfiguration
+) {
+  const newCs = { ...cs, saveCredentials: false };
+  const newStorage = { ...newCs.cloudStorage.storage };
+  // The following two lines remove the sensitive fields from the storage configuration,
+  // which should be ok, but isn't; so keep in the sensitive fields.
+  // newCs.sensitiveFieldValues = {};
+  // newStorage.configuration = {};
+  const newCloudStorage = {
+    ...newCs.cloudStorage,
+    storage: newStorage,
+  };
+  newCs.cloudStorage = newCloudStorage;
+  return newCs;
+}
+
 export function storageDefinitionFromConfig(
   config: SessionStartCloudStorageConfiguration
 ) {
