+++ tests/cypress/e2e/projectV2.spec.ts
@@ -813,8 +813,18 @@ describe("launch sessions with cloud storage", () => {
         const csConfig = req.body.cloudstorage;
         expect(csConfig.length).equal(1);
         const storage = csConfig[0];
-        expect(storage.configuration).to.not.have.property("access_key_id");
-        expect(storage.configuration).to.not.have.property("secret_access_key");
+        // The following two lines are for when the credentials are not in the config, but
+        // this causes problems, allow them to be there for now
+        // See also projectCloudStorage.utils.ts:storageDefinitionAfterSavingCredentialsFromConfig
+        // expect(storage.configuration).to.not.have.property("access_key_id");
+        // expect(storage.configuration).to.not.have.property("secret_access_key");
+
+        expect(storage.configuration).to.have.property("access_key_id");
+        expect(storage.configuration).to.have.property("secret_access_key");
+        expect(storage.configuration["access_key_id"]).to.equal("access key");
+        expect(storage.configuration["secret_access_key"]).to.equal(
+          "secret key"
+        );
         req.reply({ body: sessions[0] });
       }).as("createSession");
     });
